Move kernel symbolization logic into BlazeSymbolizer
Move the kernel symbolization logic out of ResolverMap into
BlazeSymbolizer so that we can soon retire ResolverMap.
